+++ src/components/index.ts
@@ -1,4 +1,3 @@
 export * from './common';
-export * from './types';
 export * from './product';
 export * from './ecosystem';